Refactor: ssg on shared page
