+++ next.config.mjs
@@ -2,7 +2,7 @@
 const nextConfig = {
   reactStrictMode: true,
   images: {
-    domains: ['jasonwatmore.com', 'codeit-images.codeit.com', 'codeit-frontend.codeit.com', "reactjs.org", "assets.vercel.com", "tanstack.com", "storybook.js.org", "testing-library.com", "static.cdninstagram.com", "s.pstatic.net", "static.cdninstagram.com", "s.pstatic.net"],
+    domains: ['jasonwatmore.com', 'codeit-images.codeit.com', 'codeit-frontend.codeit.com', "reactjs.org", "assets.vercel.com", "tanstack.com", "storybook.js.org", "testing-library.com", "static.cdninstagram.com", "s.pstatic.net", "static.cdninstagram.com", "s.pstatic.net", "ssl.pstatic.net", "data1.pokemonkorea.co.kr",],
   },
 
 };
